api/requirements.txt
fastapi==0.96.0
uvicorn==0.22.0
tensorflow==2.11.0
Pillow==9.5.0
opencv-python==4.7.0.72
python-multipart==0.0.6
flwr==1.4.0
websockets==11.0.3


